Move unit tests to Android module
