+++ packages/google_maps_flutter/google_maps_flutter/android/build.gradle
@@ -38,13 +38,16 @@ android {
         androidTestImplementation 'androidx.test:runner:1.2.0'
         androidTestImplementation 'androidx.test:rules:1.2.0'
         androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+        testImplementation 'junit:junit:4.12'
+        testImplementation 'org.mockito:mockito-core:3.2.4'
+        testImplementation 'androidx.test:core:1.2.0'
+        testImplementation "org.robolectric:robolectric:4.3.1"
     }
 
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
-
 
     testOptions {
         unitTests.includeAndroidResources = true
@@ -57,9 +60,4 @@ android {
             }
         }
     }
-}
-
-dependencies {
-    testImplementation 'junit:junit:4.12'
-    testImplementation 'org.mockito:mockito-core:3.2.4'
 }
