+++ packages/image_picker/image_picker/android/build.gradle
@@ -35,8 +35,17 @@ android {
         implementation 'androidx.core:core:1.0.2'
         implementation 'androidx.annotation:annotation:1.0.0'
         implementation 'androidx.exifinterface:exifinterface:1.3.0'
+
+        testImplementation 'junit:junit:4.12'
+        testImplementation 'org.mockito:mockito-core:3.10.0'
+        testImplementation 'androidx.test:core:1.2.0'
+        testImplementation "org.robolectric:robolectric:4.3.1"
     }
 
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 
     testOptions {
         unitTests.includeAndroidResources = true
