+++ packages/image_picker/image_picker/android/src/test/resources/pngImage.png
